Added a Curtain class to allow bespoke scene transitions
Transition is a subclass of Curtain that uses an Animate file as before. The StageManager calls the open/close on the Curtain class and Curtain class calls back when its open/close process is finished. Haven't totally figured out the best way to call the old Transition vs. the new Curtain; right now if there's a curtain on the game call, it'll do that.